feat: repository 기능 변경으로 인한 mock repository 변경
